# AccountMs/src/main/resources/openApiAccount.yaml
openapi: 3.0.1
info:
  title: MicroServicio Banco XYZ
  description: API para gestionar cuentas bancarias de clientes
  version: 1.0.0
servers:
  - url: http://localhost:8082
    description: Servidor Local para AccountMs

paths:
  /cuentas:
    get:
      summary: Listar todas las cuentas bancarias
      operationId: listAllAccounts
      tags:
        - Cuentas
      responses:
        '200':
          description: Lista de cuentas bancarias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CuentaResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Crear una nueva cuenta bancaria para un cliente
      operationId: createAccount
      tags:
        - Cuentas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaRequest'
      responses:
        '201':
          description: Cuenta bancaria creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuentaResponse'
        '400':
          description: Solicitud inválida (saldo inicial incorrecto o cliente inexistente)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cuentas/{id}:
    get:
      summary: Obtener detalles de una cuenta bancaria por su ID
      operationId: getAccountById
      tags:
        - Cuentas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cuenta bancaria encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CuentaResponse'
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      summary: Eliminar una cuenta bancaria por su ID
      operationId: deleteAccountById
      tags:
        - Cuentas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Cuenta eliminada exitosamente
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cuentas/{cuentaId}/depositar:
    put:
      summary: Realizar un depósito en una cuenta bancaria
      operationId: depositToAccount
      tags:
        - Cuentas
      parameters:
        - name: cuentaId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaRequest'
      responses:
        '200':
          description: Depósito realizado exitosamente
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /cuentas/{cuentaId}/retirar:
    put:
      summary: Realizar un retiro de una cuenta bancaria
      operationId: withdrawFromAccount
      tags:
        - Cuentas
      parameters:
        - name: cuentaId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CuentaRequest'
      responses:
        '200':
          description: Retiro realizado exitosamente
        '400':
          description: Retiro inválido (fondos insuficientes)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Cuenta no encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CuentaRequest:
      type: object
      properties:
        numeroCuenta:
          type: string
        saldo:
          type: number
          format: double
          default: 0.0
        tipoCuenta:
          type: string
          enum:
            - AHORROS
            - CORRIENTE
        clienteId:
          type: integer
      required:
        - numeroCuenta
        - saldo
        - tipoCuenta
        - clienteId

    CuentaResponse:
      type: object
      properties:
        id:
          type: integer
        numeroCuenta:
          type: string
        saldo:
          type: number
          format: double
        tipoCuenta:
          type: string
          enum:
            - AHORROS
            - CORRIENTE
        clienteId:
          type: integer
      required:
        - id
        - numeroCuenta
        - saldo
        - tipoCuenta
        - clienteId

    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        details:
          type: string
      required:
        - code
        - message